refactor: split out categories to `categories.yml`
