+++ resources/categories.yml
@@ -1,0 +1,133 @@
+- key: code_editor
+  name: Code Editors & IDEs
+  description: Tools for writing and editing code, offering syntax highlighting, debugging, etc.
+  emoji: 👾
+- key: development
+  name: Development Tools
+  description: Tools aiding software development, management and distribution.
+  emoji: 💭
+- key: library
+  name: Libraries
+  description: Software libraries that implement and expose the Catppuccin palette in a programming language.
+  emoji: 📚
+- key: cli
+  name: CLI Tools
+  description: Command-line interface utilities that perform specific tasks, ranging from file management to system monitoring.
+  emoji: 🐚
+- key: terminal
+  name: Terminals
+  description: Programs providing text-based interfaces to interact with the operating system.
+  emoji: 🌱
+- key: system
+  name: System
+  description: Applications that are close to the operating system.
+  emoji: 🔧
+- key: boot_loader
+  name: Boot Loaders
+  description: Tools that are responsible for booting the computer.
+  emoji: 👢
+- key: notification_daemon
+  name: Notification Daemons
+  description: Applications that notify the user of desktop events in an asynchronous manner.
+  emoji: 😈
+- key: desktop_environment
+  name: Desktop Environments
+  description: Graphical User Interfaces providing a complete experience for interacting with the operating system.
+  emoji: 🖥️
+- key: window_manager
+  name: Window Managers
+  description: Applications that control the placement and appearance of windows in a GUI.
+  emoji: 🪟
+- key: file_manager
+  name: File Managers
+  description: Applications that help users navigate, organize and edit files/directories on a computer.
+  emoji: 🗂️
+- key: application_launcher
+  name: Application Launchers
+  description: Tools that provide a quick and easy way for users to launch applications.
+  emoji: 🗃️
+- key: browser
+  name: Browsers
+  description: Applications for accessing and navigating the World Wide Web.
+  emoji: 🏄
+- key: browser_extension
+  name: Browser Extensions
+  description: Add-ons which extend the functionality of web browsers, enhancing the browsing experience.
+  emoji: 🧩
+- key: search_engine
+  name: Search Engines
+  description: Websites that enable users to search and retrieve information from the World Wide Web.
+  emoji: 🔎
+- key: translation_tool
+  name: Translation Tools
+  description: Applications that help bridge language gaps by translating one language to another.
+  emoji: 🗺️
+- key: wiki
+  name: Wikis
+  description: Websites that hold information edited by multiple users.
+  emoji: 🧠
+- key: photo_and_video
+  name: Photo & Video
+  description: Applications that help in capturing, editing, managing, storing, or sharing photos and videos.
+  emoji: 📸
+- key: music
+  name: Music
+  description: Applications related to discovering, listening to and playing music.
+  emoji: 🎵
+- key: productivity
+  name: Productivity
+  description: Applications that make certain tasks or activities more organised and efficient.
+  emoji: 📖
+- key: email_client
+  name: Email Clients
+  description: Applications related to sending and receiving email.
+  emoji: 💌
+- key: game_development
+  name: Game Development
+  description: Applications that are designed for creating, designing and building video games.
+  emoji: 🕹️
+- key: 3d_modelling
+  name: 3D Modelling
+  description: Applications that are designed to create and manipulate objects in three-dimensional space.
+  emoji: 🖼️
+- key: note_taking
+  name: Note Taking
+  description: Tools for creating, organizing, and managing digital notes.
+  emoji: 📝
+- key: education
+  name: Education
+  description: Applications that faciliate learning, knowledge acquisition and skill development.
+  emoji: 🎓
+- key: document_viewer
+  name: Document Viewers
+  description: Applications that allow users to open, view and interact with documents.
+  emoji: 📃
+- key: social_networking
+  name: Social Networking
+  description: Applications that bring people together through text, voice, photo and/or video.
+  emoji: ✨
+- key: discussion_forum
+  name: Discussion Forums
+  description: Platforms that allow people to engage in conversation, share information, and discuss various topics.
+  emoji: 🗣️
+- key: entertainment
+  name: Entertainment
+  description: Applications designed to entertain the user.
+  emoji: 🌈
+- key: game
+  name: Games
+  description: Applications related to gaming and the idea of game-ification.
+  emoji: 🎮
+- key: analytics
+  name: Analytics
+  description: Applications related to visualizations and analytics of data.
+  emoji: 📊
+- key: artificial_intelligence
+  name: Artificial Intelligence
+  description: Applications related to machine learning, large language models, and neural networks.
+  emoji: 🤖
+- key: health_and_fitness
+  name: Health & Fitness
+  description: Applications related to healthy living, fitness, and managing stress.
+  emoji: 🏃
+# yaml-language-server: $schema=./categories.schema.json
