+++ resources/deno.lock
@@ -1,5 +1,45 @@
 {
-  "version": "3",
+  "version": "4",
+  "specifiers": {
+    "jsr:@std/path@^1.0.8": "1.0.8",
+    "jsr:@std/yaml@1.0.5": "1.0.5",
+    "npm:ajv@8.17.1": "8.17.1",
+    "npm:type-fest@4.37.0": "4.37.0"
+  },
+  "jsr": {
+    "@std/path@1.0.8": {
+      "integrity": "548fa456bb6a04d3c1a1e7477986b6cffbce95102d0bb447c67c4ee70e0364be"
+    },
+    "@std/yaml@1.0.5": {
+      "integrity": "71ba3d334305ee2149391931508b2c293a8490f94a337eef3a09cade1a2a2742"
+    }
+  },
+  "npm": {
+    "ajv@8.17.1": {
+      "integrity": "sha512-B/gBuNg5SiMTrPkC+A2+cW0RszwxYmn6VYxB/inlBStS5nx6xHIt/ehKRhIMhqusl7a8LjQoZnjCs5vhwxOQ1g==",
+      "dependencies": [
+        "fast-deep-equal",
+        "fast-uri",
+        "json-schema-traverse",
+        "require-from-string"
+      ]
+    },
+    "fast-deep-equal@3.1.3": {
+      "integrity": "sha512-f3qQ9oQy9j2AhBe/H9VC91wLmKBCCU/gDOnKNAYG5hswO7BLKj09Hc5HYNz9cGI++xlpDCIgDaitVs03ATR84Q=="
+    },
+    "fast-uri@3.0.6": {
+      "integrity": "sha512-Atfo14OibSv5wAp4VWNsFYE1AchQRTv9cBGWET4pZWHzYshFSS9NQI6I57rdKn9croWVMbYFbLhJ+yJvmZIIHw=="
+    },
+    "json-schema-traverse@1.0.0": {
+      "integrity": "sha512-NM8/P9n3XjXhIZn1lLhkFaACTOURQXjWhV4BA/RnOv8xvgqtqpAX9IO4mRQxSx1Rlo4tqzeqb0sOlruaOy3dug=="
+    },
+    "require-from-string@2.0.2": {
+      "integrity": "sha512-Xf0nWe6RseziFMu+Ap9biiUbmplq6S9/p+7w7YXP/JBHhrUDDUhwa+vANyubuqfZWTveU//DYVGsDG7RKL/vEw=="
+    },
+    "type-fest@4.37.0": {
+      "integrity": "sha512-S/5/0kFftkq27FPNye0XM1e2NsnoD/3FS+pBmbjmmtLT6I+i344KoOf7pvXreaFsDamWeaJX55nczA1m5PsBDg=="
+    }
+  },
   "remote": {
     "https://deno.land/std@0.207.0/assert/assert.ts": "9a97dad6d98c238938e7540736b826440ad8c1c1e54430ca4c4e623e585607ee",
     "https://deno.land/std@0.207.0/assert/assertion_error.ts": "4d0bde9b374dfbcbe8ac23f54f567b77024fb67dbb1906a852d67fe050d42f56",
@@ -117,6 +157,8 @@
     "https://deno.land/x/catppuccin_toolbox@deno-lib-v1.0.0/deno-lib/validateYaml.ts": "aa26af13d5d0b1cddc1c2df5f025ce8666b44070be29db063e336e76779402a4",
     "https://esm.sh/v135/ajv@8.12.0": "cc1a73af661466c7f4e6a94d93ece78542d700f2165bdb16a531e9db8856c5aa",
     "https://esm.sh/v135/ajv@8.12.0/denonext/ajv.mjs": "4645df9093d0f8be0e964070a4a7aea8adea06e8883660340931f7a3f979fc65",
+    "https://esm.sh/v135/ajv@8.12.0/denonext/dist/ajv.js": "1c8812d2f11bfee54eb884b5d006b4b3a5c5d3d8f43a170c4c18adfca50aa25e",
+    "https://esm.sh/v135/ajv@8.12.0/dist/ajv.js": "ebdeb904c7f7bc183d4b165988e5f231110c0806d2ee5b76c1e96ec0e2564b55",
     "https://esm.sh/v135/fast-deep-equal@3.1.3/denonext/fast-deep-equal.mjs": "6313b3e05436550e1c0aeb2a282206b9b8d9213b4c6f247964dd7bb4835fb9e5",
     "https://esm.sh/v135/json-schema-traverse@1.0.0/denonext/json-schema-traverse.mjs": "c5da8353bc014e49ebbb1a2c0162d29969a14c325da19644e511f96ba670cc45",
     "https://esm.sh/v135/type-fest@4.8.3": "fabf6997a3adf11e45d11e3c4d58d0634ab98f9d72f3116f3694123818ccd457",
@@ -125,6 +167,15 @@
     "https://esm.sh/v135/type-fest@4.8.3/source/simplify.d": "6d148e26b1e58ccc7aa9699e2d0de830f0bc4221b6b098700763fd757ff7b5ad",
     "https://esm.sh/v135/type-fest@4.8.3/source/union-to-intersection.d": "e83c6075c4a23e3b01dd9ff87e31a0d91b7a5fdf17bf606d82c9a72e895d939d",
     "https://esm.sh/v135/uri-js@4.4.1/denonext/uri-js.mjs": "901d462f9db207376b39ec603d841d87e6b9e9568ce97dfaab12aa77d0f99f74",
+    "https://raw.githubusercontent.com/catppuccin/toolbox/refs/heads/feat/deno-lib-multiple-schemas/deno-lib/mod.ts": "c72687f31c7267d121441bb872dab08621a679f6b46e59355391b11f8f9dc160",
+    "https://raw.githubusercontent.com/catppuccin/toolbox/refs/heads/feat/deno-lib-multiple-schemas/deno-lib/updateReadme.ts": "e349b1c0b2c1d1fc317bbc5b506f733d3c03d48ef26d0d62eafb7cc5141ed492",
+    "https://raw.githubusercontent.com/catppuccin/toolbox/refs/heads/feat/deno-lib-multiple-schemas/deno-lib/validateYaml.ts": "b0059c257e4b1ed2471080fa3ae4579e1c9a8eded8e263c61d91dac92101c420",
     "https://raw.githubusercontent.com/catppuccin/userstyles/6ca6c437362af2b221c1af30e3c88119020a8ac3/scripts/userstyles.schema.json": "9a543a3da71cf668ce80731dc417b8ac746d0ec20bf255ce93d1541214f34fcc"
+  },
+  "workspace": {
+    "dependencies": [
+      "jsr:@std/path@^1.0.8",
+      "npm:type-fest@4.37.0"
+    ]
   }
 }
