+++ resources/ports.schema.json
@@ -171,47 +171,6 @@
         }
       }
     },
-    "categories": {
-      "$id": "#category-definitions",
-      "title": "Category Definitions",
-      "description": "The categories listed in the README.",
-      "type": "array",
-      "items": {
-        "$id": "#category-definitions/item",
-        "title": "Category Item",
-        "type": "object",
-        "required": [
-          "key",
-          "name",
-          "description",
-          "emoji"
-        ],
-        "additionalProperties": false,
-        "properties": {
-          "key": {
-            "$id": "#category-definitions/category/key",
-            "title": "key",
-            "type": "string",
-            "$ref": "#/$defs/category"
-          },
-          "name": {
-            "$id": "#category-definitions/category/name",
-            "title": "Name",
-            "type": "string"
-          },
-          "description": {
-            "$id": "#category-definitions/category/description",
-            "title": "Description",
-            "type": "string"
-          },
-          "emoji": {
-            "$id": "#category-definitions/category/emoji",
-            "title": "Emoji",
-            "type": "string"
-          }
-        }
-      }
-    },
     "showcases": {
       "$id": "#showcases",
       "title": "Showcases",
@@ -278,7 +237,12 @@
       "$id": "#port-categories",
       "title": "Categories",
       "description": "The categories that fit the port the most, the first category is the primary category which the port will be listed under on the README.",
-      "$ref": "#/$defs/categories"
+      "type": "array",
+      "minItems": 1,
+      "maxItems": 3,
+      "items": {
+        "$ref": "categories.schema.json#/$defs/category"
+      }
     },
     "port-platform": {
       "$id": "#port-platform",
@@ -316,7 +280,9 @@
           "type": "array",
           "items": {
             "type": "string",
-            "enum": ["agnostic"]
+            "enum": [
+              "agnostic"
+            ]
           },
           "minItems": 1,
           "maxItems": 1
@@ -357,56 +323,6 @@
       ],
       "examples": [
         "pink"
-      ]
-    },
-    "categories": {
-      "$id": "#categories",
-      "type": "array",
-      "minItems": 1,
-      "maxItems": 3,
-      "items": {
-        "$ref": "#/$defs/category"
-      }
-    },
-    "category": {
-      "$id": "#category",
-      "type": "string",
-      "enum": [
-        "3d_modelling",
-        "analytics",
-        "application_launcher",
-        "artificial_intelligence",
-        "boot_loader",
-        "browser",
-        "browser_extension",
-        "cli",
-        "code_editor",
-        "desktop_environment",
-        "development",
-        "discussion_forum",
-        "document_viewer",
-        "education",
-        "email_client",
-        "entertainment",
-        "file_manager",
-        "game",
-        "game_development",
-        "health_and_fitness",
-        "library",
-        "music",
-        "note_taking",
-        "notification_daemon",
-        "photo_and_video",
-        "productivity",
-        "search_engine",
-        "self_hosted",
-        "social_networking",
-        "system",
-        "terminal",
-        "translation_tool",
-        "userstyle",
-        "wiki",
-        "window_manager"
       ]
     }
   }
