+++ resources/types/mod.ts
@@ -1,4 +1,5 @@
 export * as PortsSchema from "@/types/ports.d.ts";
+export * as CategoriesSchema from "@/types/categories.d.ts";
 export * as UserStylesSchema from "catppuccin-userstyles/scripts/types/userstyles.d.ts";
 
 import { Entries } from "type-fest";
